Setup Timeline, Interactable, and Audio
I set up the timeline so the game now finishes after the player gets 3 interactable objects into the box. Background music is now audible. Finally, all objects are interactable with all but the book being wrong and getting reset. The yellow book, only interactable book, is one of the correct items and will score a point.